+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.2 - 2017-08-03
+### Changed
+- Added a `warning` (orange) color to **Button**.
+- Added `onUpdate` to **Drawer**.
+- Removed `ignorePadding` from **Drawer**.
+
 ## 1.1.1 - 2017-08-03
 ### Changed
 - Upgraded extract-text-webpack-plugin.